refactor: moved files to appropriate dirs (#42)
* ♻️ Moved auth views to auth dir

* style: fixed sider issues

* style: formatted code
